Move `isGoogleDocsWrapper` to `slate-editor`
