+++ packages/slate-commons/src/index.ts
@@ -1,11 +1,5 @@
 export * as EditorCommands from './commands';
-export {
-    createDeserializeElement,
-    isGoogleDocsWrapper,
-    nodeIdManager,
-    useSavedSelection,
-    withoutNodes,
-} from './lib';
+export { createDeserializeElement, nodeIdManager, useSavedSelection, withoutNodes } from './lib';
 export * from './constants';
 export * from './plugins';
 export * as Selection from './selection';
