+++ packages/slate-commons/src/lib/index.ts
@@ -1,5 +1,4 @@
 export { createDeserializeElement } from './createDeserializeElement';
-export { isGoogleDocsWrapper } from './isGoogleDocsWrapper';
 export { nodeIdManager } from './nodeIdManager';
 export { useSavedSelection } from './useSavedSelection';
 export { withoutNodes } from './withoutNodes';
