+++ packages/slate-editor/src/lib/isGoogleDocsWrapper.ts
